# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Homework Week 9",
    "description": "Restfull API and Middleware",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
      "ErrorNotFound": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "error not found"
          }
        }
      },
      "InvalidCredentials": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "invalid email or password"
          }
        }
      },
...
      "ErrorInteral": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Unauthenticated"
          }
        }
      },
...
      "ReqBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "example@mail.com"
          },
          "password": {
            "type": "string",
            "example": "admin"
          },
          "gender": {
            "type": "string",
            "example": "male"
          },
          "role": {
            "type": "string",
            "example": "admin"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "create a new user",
        "requestBody": {
          "description": "Insert email, password, and role",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReqBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "user created successfully"
                    },
                    "email": {
                      "type": "string",
                      "example": "example@mail.com"
                    },
                    "gender": {
                      "type": "string",
                      "example": "male"
                    },
                    "role": {
                      "type": "string",
                      "example": "admin"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInteral"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "example@mail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "admin"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "login success"
                    },
                    "accessToken": {
                      "type": "string",
                      "example": "generated token"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidCredentials"
                }
              }
            }
          },
          "404": {
            "description": "Error Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorNotFound"
                }
              }
            }
          }
        }
      }
    },
...
  }
}